feat: auto generate config from previous version
