+++ libs/@local/eslint/package.json
@@ -4,11 +4,16 @@
   "private": true,
   "description": "Shared ESLint configuration",
   "type": "module",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
+  "exports": {
+    ".": {
+      "default": "./index.js",
+      "types": "./index.d.ts"
+    },
+    "./deprecated": {
+      "default": "./deprecated/index.js",
+      "types": "./deprecated/index.d.ts"
+    }
+  },
   "scripts": {
     "build": "rimraf dist && tsc --build tsconfig.build.json"
   },
@@ -17,13 +22,14 @@
     "@eslint/eslintrc": "3.2.0",
     "effect": "3.11.0",
     "eslint": "9.16.0",
+    "eslint-config-flat-gitignore": "0.3.0",
     "eslint-config-sheriff": "25.3.0",
     "eslint-plugin-airbnb": "0.0.1-security",
+    "eslint-plugin-canonical": "5.0.0",
     "eslint-plugin-react-hooks": "5.1.0",
     "eslint-plugin-storybook": "0.11.1",
     "eslint-unicorn": "55.0.0",
-    "globals": "15.13.0",
-    "typescript-eslint": "8.17.0"
+    "globals": "15.13.0"
   },
   "devDependencies": {
     "@types/eslint__eslintrc": "2.1.2",
