+++ libs/@local/eslint/src/deprecated/block.ts
@@ -1,0 +1,42 @@
+import getGitignorePatterns from "eslint-config-flat-gitignore";
+import typescriptEslint from "@typescript-eslint/eslint-plugin";
+// @ts-expect-error - react-hooks does not expose types
+import reactHooks from "eslint-plugin-react-hooks";
+import react from "eslint-plugin-react";
+import unicorn from "eslint-plugin-unicorn";
+import { fixupPluginRules } from "@eslint/compat";
+import { ignores } from "eslint-config-sheriff";
+
+export const create = () => [
+  {
+    plugins: {
+      "@typescript-eslint": typescriptEslint,
+      "react-hooks": fixupPluginRules(reactHooks),
+      react,
+      unicorn,
+    },
+
+    rules: {
+      curly: ["error", "multi-line"],
+
+      "import/no-extraneous-dependencies": [
+        "error",
+        {
+          devDependencies: true,
+        },
+      ],
+
+      "jsx-a11y/label-has-associated-control": "off",
+      "react-hooks/exhaustive-deps": "error",
+      "react-hooks/rules-of-hooks": "error",
+      "react/jsx-key": "error",
+      "react/jsx-no-useless-fragment": "error",
+      "react/no-danger": "error",
+      "react/self-closing-comp": "error",
+    },
+  },
+  getGitignorePatterns({ strict: false }),
+  {
+    ignores,
+  },
+];
