--- libs/@local/eslint/src/deprecated/config.ts
@@ -1,5 +1,0 @@
-export interface DeprecatedConfig {
-  pathsOverrides?: {
-    tsconfigLocation: string;
-  };
-}
