+++ libs/@local/eslint/src/deprecated/disable.ts
@@ -1,0 +1,12 @@
+import { Array, pipe, Record } from "effect";
+import { ESConfig } from "../utils.js";
+
+export const disableRules = (rules: readonly string[]): readonly ESConfig[] =>
+  process.env.CHECK_TEMPORARILY_DISABLED_RULES === "true"
+    ? []
+    : pipe(
+        rules,
+        Array.map((rule) => [rule, "off"] as const),
+        Record.fromEntries,
+        (rules) => [{ rules }],
+      );
